Moved proxies from Effects to Proxies. Shader grabber to Components.
