+++ CCL.Types/Proxies/InternalExternalSnapshotSwitcherProxy.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace CCL.Types.Effects
+namespace CCL.Types.Proxies
 {
     public class InternalExternalSnapshotSwitcherProxy : MonoBehaviour
     {
