+++ CCL.Types/Proxies/PlayerDistanceGameObjectsDisablerProxy.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace CCL.Types.Effects
+namespace CCL.Types.Proxies
 {
     public class PlayerDistanceGameObjectsDisablerProxy : MonoBehaviour
     {
